Move depth iteration to the right module
