+++ src/arena/mod.rs
@@ -7,7 +7,7 @@ pub mod directed;
 pub mod iterables;
 mod utils;
 
-pub use depth::{DepthFirstAccumulation, DepthFirstArenaTree};
-pub use directed::{ArenaNode, BreadthFirstIterator, DepthFirstIterator, DirectedArenaTree};
+pub use depth::{DepthFirstAccumulation, DepthFirstArenaTree, DepthFirstIterator};
+pub use directed::{ArenaNode, BreadthFirstIterator, DirectedArenaTree};
 use iterables::BaseDirectionIterable;
 pub use iterables::{BreadthFirstIterable, DepthFirstIterable, DirectionIterable, Nodelike};
